+++ src/main/java/org/example/calendar_advanced/global/config/WebMvcConfiguration.java
@@ -1,8 +1,6 @@
 package org.example.calendar_advanced.global.config;
 
-import org.example.calendar_advanced.global.filter.CommentFilter;
-import org.example.calendar_advanced.global.filter.ScheduleFilter;
-import org.example.calendar_advanced.global.filter.UserFilter;
+import org.example.calendar_advanced.global.filter.AuthFilter;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -12,29 +10,15 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 public class WebMvcConfiguration implements WebMvcConfigurer {
 
     @Bean
-    public FilterRegistrationBean userFilter(){
-        FilterRegistrationBean registrationBean = new FilterRegistrationBean(new UserFilter());
+    public FilterRegistrationBean authFilter(){
+        FilterRegistrationBean registrationBean = new FilterRegistrationBean(new AuthFilter());
         registrationBean.setOrder(1);
+        registrationBean.addUrlPatterns("/api/login");
         registrationBean.addUrlPatterns("/api/users/*");
-
-        return registrationBean;
-    }
-
-    @Bean
-    public FilterRegistrationBean scheduleFilter(){
-        FilterRegistrationBean registrationBean = new FilterRegistrationBean(new ScheduleFilter());
-        registrationBean.setOrder(1);
         registrationBean.addUrlPatterns("/api/schedules/*");
-
-        return registrationBean;
-    }
-
-    @Bean
-    public FilterRegistrationBean commentFilter(){
-        FilterRegistrationBean registrationBean = new FilterRegistrationBean(new CommentFilter());
-        registrationBean.setOrder(1);
         registrationBean.addUrlPatterns("/api/comments/*");
 
         return registrationBean;
     }
+
 }
